requirements.txt
absl-py==1.3.0
altair==4.2.0
anyascii==0.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
arviz==0.16.1
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.6.2
branca==0.5.0
cachetools==5.2.0
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
commonmark==0.9.1
contourpy==1.1.1
contractions==0.1.73
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
docutils==0.19
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==1.0.0
fake-useragent==0.1.11
fastjsonschema==2.16.3
filelock==3.12.4
findspark==2.0.1
Fiona==1.8.21
Flask==2.3.2
flatbuffers==22.10.26
folium==0.12.1.post1
fonttools==4.43.1
fqdn==1.5.1
future==0.18.2
gast==0.4.0
geojson==2.5.0
geopandas==0.11.1
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.50.0
h11==0.14.0
h5netcdf==1.2.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==6.1.0
install==1.3.5
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.2
jsonpointer==2.3
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.10.0
Keras-Preprocessing==1.1.2
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
kotlin-jupyter-kernel==0.11.0.385
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
metakernel==0.29.5
mistune==2.0.5
munch==2.5.0
mysqlclient==2.2.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.5
networkx==3.2
nltk==3.7
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.3
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
osmnet==0.1.6
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.10.0
plotly==5.11.0
prometheus-client==0.16.0
prompt-toolkit==3.0.31
protobuf==4.24.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyahocorasick==1.4.4
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0b4
pygeos==0.13
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.4.0
pyrsistent==0.18.1
PySocks==1.7.1
pyspark==3.4.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
Rtree==1.0.0
scikit-learn==1.1.3
scipy==1.9.1
seaborn==0.12.1
selenium==4.5.0
semver==2.13.0
Send2Trash==1.8.0
Shapely==1.8.4
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spylon==0.3.0
spylon-kernel==0.4.1
stack-data==0.5.0
statsmodels==0.14.0
streamlit==1.27.2
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.17.1
textsearch==0.0.24
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.12
validators==0.20.0
virtualenv==20.24.5
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.7
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
xarray==2023.9.0
xarray-einstats==0.6.0
xgboost==1.7.2
zipp==3.9.0
